auto-luks: require new `mountPoint` option
With the upcoming systemd v242 change in nixpkgs there will be a
breaking change to how auto-luks used to work. We are going to remove a
commit from the systemd fork that was added for auto-luks to work. The
auto-luks feature will continue to work but another mount option is
required to mark those `fileSystems` as not local. All mount points that
carry the `_netdev` mount option are not part of the `local-fs.target`
and thus not blocking the start of things like the sshd daemon.

Adding this option is (so far) the least intrusive way to ensure that
systems will not break. Every user of the feature will be required to
att only one more value to it's configuration.

Trying to automatically detect which mount points are mounted from the
unlocked block devices is not really feasible in all/most cases. There
might be another layer of indirection in between the actual (luks)
device and the mount point. Good examples that will make it hard for us
to guess are things like RAID, LVM, ZFS, ….

Dropping the systemd patch removes a bunch of issues and race conditions
in regards to creating `/run/<service>` and `/var/lib/<service>`
directories which can be handled by systemd just fine. This is also part
of the ongoing effort to remove large a portion of our `PreStart`
scripts that just create directories. In the long run there should be
less custom scripting required to get a service with (temporary and
persistent) state up and running.